Create functions to perform operations
When the key "=" is pressed, the conditional within addValue will
call the operation() function, which has a switch conditional to
evaluate what operation to perfomr according to the class received
by the key that provided the operator.

There are 4 cases in the switch conditional, which are:

*add
*subtract
*multiply
*div

Each of these functions displays the correct value on the calculator;
however, when an extra symbol is pressed and erased the result
yields NaN. This problem will be addressed from here on.