Correct backspace problem and use floats
When a second operator was added and the removed with backspace the
calculator returned NaN. The problem was addressed by making the program
to yield the result either by pressing the equal key or another symbol.
However, when pressing another symbol, the calculator is ready to
compute another operation.

This was achievied by means of a counter. When it detected there were
two symbols on the display the same logic used with the equal symbol
was used, but the display object was summoned at the end to show
the symbol on the calculator after the result of the previous operation.

The calculator now is able to use floats, which was achieved by
replacing parseInt with parseFloat.