Limit the use of the key point
A counter was declared to measure the number of times the key point
has been pressed. If it is greater or equal to 2, then the calculator
won't show another point.

When the CE key is pressed, it erases everything on the display and
reset all the variables, including the counter for the point.

In the backspace function the element to remove was stored in a
variable, so the system could compare if it is a point. Given this case,
then the counter for the point is reset.