Link KeyPad for numbers and operations
By means of an event listener to the window a new function is summoned,
which was named addValueKeyPad(). This function has the same logic
as addValue(), which was renamed as addValueClick(). However, at the
outset the constant key was declared to select the HTML element that
matches with the key code, providing the function with the required
classes to work. If key is empty, then the function returns nothing.

The key for the backspace remains to be added.